text2digit library error handling
converting an int to an int is a redundant operation. Created error handling to identify digits in english text input. We only want english text in the input to prevent the user from accidentally putting in decimal, binary, octal, etc. into the input when the input type is English Text.